lsd_slam viewer built on El Captain
- additionally when running rosrun:
$ export
DYLD_LIBRARY_PATH=DYLD_LIBRARY_PATH:$(pwd)/devel/lib:/Users/felixschalle
r/ros_catkin_ws/install_isolated/lib

 to set the path to the dylib libraries in the rosrun shell script.

+++ lsd_slam/lsd_slam_viewer/src/PointCloudViewer.h
@@ -2,7 +2,7 @@
 * This file is part of LSD-SLAM.
 *
 * Copyright 2013 Jakob Engel <engelj at in dot tum dot de> (Technical University of Munich)
-* For more information see <http://vision.in.tum.de/lsdslam> 
+* For more information see <http://vision.in.tum.de/lsdslam>
 *
 * LSD-SLAM is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
@@ -132,7 +132,7 @@ public:
 		int showCurrentCam_i = showCurrentCam;
 		int isFix_i = isFix;
 
-		float x,y,z;
+		qreal x,y,z;
 		frame.getPosition(x,y,z);
 
     	snprintf(buf, 1000, "Animation: %d at %lf (dur %lf) S: %f %f %d %d %d %d %d Frame: %lf %lf %lf %lf %f %f %f %d",
@@ -224,5 +224,3 @@ private:
 
 	void remakeAnimation();
 };
-
-
